Feature/rizqiaryansa/dog fooding download manager feedback test

+++ features/common/feedback_form/src/main/java/com/tokopedia/feedback_form/feedbackpage/domain/response/FeedbackFormResponse.kt
@@ -11,7 +11,7 @@ data class FeedbackFormResponse(
 
 data class Data(
         @SerializedName("feedbackID")
-        var feedbackID: Int = 0,
+        var feedbackID: Long = 0,
         @SerializedName("message")
         var message: String = ""
 )
