+++ features/common/feedback_form/src/main/java/com/tokopedia/feedback_form/feedbackpage/ui/feedbackpage/FeedbackPageContract.kt
@@ -16,7 +16,7 @@ interface FeedbackPageContract {
         fun showLoadingFrame()
         fun hideLoadingFrame()
         fun setSubmitFlag()
-        fun checkUriImage(feedbackId: Int, imageCount: Int)
+        fun checkUriImage(feedbackId: Long, imageCount: Int)
         fun goToTicketCreatedActivity(issueUrl: String?)
         fun showError(throwable: Throwable)
         fun setFeedbackData(model: FeedbackModel)
@@ -26,8 +26,8 @@ interface FeedbackPageContract {
     interface Presenter{
         fun getFeedbackData()
         fun sendFeedbackForm(feedbackFormRequest: FeedbackFormRequest)
-        fun sendAttachment(feedbackId: Int, fileData: MultipartBody.Part, totalImage: Int, imageCount: Int)
-        fun commitData(feedbackId: Int)
+        fun sendAttachment(feedbackId: Long, fileData: MultipartBody.Part, totalImage: Int, imageCount: Int)
+        fun commitData(feedbackId: Long)
         fun getImageList(selectedImage: ArrayList<String>): MutableList<BaseImageFeedbackUiModel>
         fun initImageData(): MutableList<BaseImageFeedbackUiModel>
         fun screenshotImageResult(data: ScreenshotData) : MutableList<BaseImageFeedbackUiModel>
